ai-service/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-google-genai
chromadb
python-dotenv
pydantic
